sync: Move syncing tests to tests/sync.rs
Keep unit tests in src/sync/mod.rs and move everything else to
tests/sync.rs
